src/requirements/requirements-dev.txt
-r requirements.txt

django-debug-toolbar==1.4